+++ src/main/java/com/example/bookstorebg/service/OrderService.java
@@ -3,6 +3,7 @@ package com.example.bookstorebg.service;
 import com.example.bookstorebg.entity.Order;
 import com.example.bookstorebg.entity.OrderItem;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 import java.util.Map;
@@ -11,7 +12,7 @@ public interface OrderService {
 
     List<Order> getOrders(Long user_id);
 
-    void addOrder(Long user_id, Double price, String address, String receiver, String tele, Timestamp time);
+    void addOrder(Long user_id, BigDecimal price, String address, String receiver, String tele, Timestamp time);
 
     List<Order> getAllOrders();
